変更

+++ db/migrate/20220907064505_create_post_tags.rb
@@ -1,8 +1,8 @@
 class CreatePostTags < ActiveRecord::Migration[6.1]
   def change
     create_table :post_tags do |t|
-      t.references  :post, foreign_key: true, type: :integer
-      t.references  :tag,  foreign_key: true, type: :integer
+      t.references  :post, foreign_key: true, type: :bigint
+      t.references  :tag,  foreign_key: true, type: :bigint
       t.timestamps
     end
     # 同じタグを２回保存するのはできないようになる
