+++ db/migrate/20220916085754_create_entries.rb
@@ -1,8 +1,8 @@
 class CreateEntries < ActiveRecord::Migration[6.1]
   def change
     create_table :entries do |t|
-      t.references :customer, null: false, foreign_key: true, type: :integer
-      t.references :room,     null: false, foreign_key: true, type: :integer
+      t.references :customer, null: false, foreign_key: true, type: :bigint
+      t.references :room,     null: false, foreign_key: true, type: :bigint
       t.timestamps
     end
   end
