+++ db/migrate/20220916090838_create_messages.rb
@@ -1,8 +1,8 @@
 class CreateMessages < ActiveRecord::Migration[6.1]
   def change
     create_table :messages do |t|
-      t.references :customer, null: false, foreign_key: true, type: :integer
-      t.references :room,     null: false, foreign_key: true, type: :integer
+      t.references :customer, null: false, foreign_key: true, type: :bigint
+      t.references :room,     null: false, foreign_key: true, type: :bigint
       t.text       :content
       t.timestamps
     end
